# packages/spike.land/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Spike Land API
  description: |
    API for Spike Land - AI-Powered App Platform with Image Enhancement

    ## Platform Overview
    Spike Land is a platform where users create, fork, and manage AI-powered applications.
    The flagship application, Pixel, provides AI-powered image enhancement.

    ## Authentication
    All protected endpoints require a valid session cookie from NextAuth.js.
    Alternatively, you can use Bearer token authentication for API requests:
    ```
    Authorization: Bearer {session_token}
    ```

    ## Token Economy
    Users have a platform-wide token balance used for:
    - Image enhancement operations (cost depends on resolution tier)
    - Premium features in apps

    Token costs for Pixel image enhancement:
    - TIER_1K: 2 tokens
    - TIER_2K: 5 tokens
    - TIER_4K: 10 tokens

    Tokens regenerate automatically (1 token per 15 minutes, max 100).

    ## Rate Limiting
    Most endpoints are rate-limited per user. Rate limit information is included in response headers:
    - X-RateLimit-Remaining: Requests remaining in current window
    - X-RateLimit-Reset: Unix timestamp when limit resets

    Example rate limits:
    - Image enhancement: 3 requests per minute
    - Image upload: 10 requests per minute
    - Album operations: 10 requests per minute

    ## Error Handling
    The API uses standard HTTP status codes:
    - 200 OK: Successful request
    - 201 Created: Resource created
...

...

    ## Base URL
...
servers:
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie (automatically set after login)
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token for API requests
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Unauthorized"
        code:
          type: string
          description: Machine-readable error code
          example: "UNAUTHORIZED"
        details:
          type: object
          description: Additional error context
          example: { "field": "email", "reason": "invalid format" }
        title:
          type: string
          description: Short error title
          example: "Authentication Failed"
        suggestion:
          type: string
          description: Suggested action to resolve the error
          example: "Please log in again"
    ValidationError:
      type: object
      required:
        - error
        - field
      properties:
        error:
          type: string
          description: Validation error message
          example: "Invalid input"
        field:
          type: string
          description: Field that failed validation
          example: "email"
        code:
          type: string
          enum:
            - REQUIRED_FIELD
            - INVALID_FORMAT
            - FILE_TOO_LARGE
            - UNSUPPORTED_FORMAT
          description: Validation error code
          example: "INVALID_FORMAT"

    # ========================================
    # Enumerations
    # ========================================
...
  responses:
    Unauthorized:
      description: Authentication required or failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Unauthorized"
            code: "UNAUTHORIZED"
            title: "Authentication Failed"
            suggestion: "Please log in again"
...
    BadRequest:
      description: Invalid request input
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/Error"
              - $ref: "#/components/schemas/ValidationError"
          example:
            error: "Invalid input"
            code: "INVALID_INPUT"
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Requests remaining in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when limit resets
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Rate limit exceeded"
            code: "RATE_LIMIT_EXCEEDED"
            title: "Too Many Requests"
            suggestion: "Please wait before making another request"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Internal server error"
            code: "INTERNAL_SERVER_ERROR"
            title: "Server Error"
            suggestion: "Please try again later"

  # ========================================
  # Request Parameters
  # ========================================
...
paths:
...
  /api/images/batch-upload:
    post:
      summary: Upload multiple images in a batch
      tags: [Batch Operations]
      operationId: batchUploadImages
      description: |
        Upload up to 20 images in a single request. Each file is processed sequentially.
        Supports partial success - if some files fail, others are still processed.
        File size limits: 10MB per file, 50MB total batch size.
        All files are uploaded to R2 first, then database records created in transaction.
        If database transaction fails, all R2 uploads are automatically cleaned up.
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 20
                  description: Image files to upload (max 20, each max 10MB)
      responses:
        "200":
          description: Batch upload completed (partial success allowed)
          headers:
            X-Request-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required: [success, results, summary]
                properties:
                  success:
                    type: boolean
                    description: True if at least one file succeeded
                  results:
                    type: array
                    items:
                      type: object
                      required: [success, filename]
                      properties:
                        success:
                          type: boolean
                        filename:
                          type: string
                        imageId:
                          type: string
                          nullable: true
                        url:
                          type: string
                          format: uri
                          nullable: true
                        width:
                          type: integer
                          nullable: true
                        height:
                          type: integer
                          nullable: true
                        size:
                          type: integer
                          nullable: true
                        format:
                          type: string
                          nullable: true
                        error:
                          type: string
                          nullable: true
                        errorType:
                          type: string
                          enum: [validation, upload, database, unknown]
                          nullable: true
                  summary:
                    type: object
                    required: [total, successful, failed]
                    properties:
                      total:
                        type: integer
                      successful:
                        type: integer
                      failed:
                        type: integer
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalServerError"
...